Refactor to many-to-many implementation with nested fields.
Locations are attached to Connections by way of Endpoints. You are able
to create a new Location by itself, or create a new Location and attach
it to an existing Location. The new action and nested form takes care
of creating the intermediate parameters so that the whole graph is
created when @location.save is called.